Ajustado e testado pelo swagger endpoints de Usuarios
