+++ src/main/java/com/project/softwave/backend_SoftWave/config/SecurityConfiguracao.java
@@ -57,6 +57,7 @@ public class SecurityConfiguracao {
             new AntPathRequestMatcher("/h2-console/**"),
             new AntPathRequestMatcher("/h2-console/**/**"),
             new AntPathRequestMatcher("/error/**"),
+            new AntPathRequestMatcher(("/**"))
 
     };
