+++ src/main/java/com/project/softwave/backend_SoftWave/controller/UsuarioFisicoController.java
@@ -1,6 +1,8 @@
 package com.project.softwave.backend_SoftWave.controller;
 
 import com.project.softwave.backend_SoftWave.dto.*;
+import com.project.softwave.backend_SoftWave.dto.UsuarioFisico.UsuarioFisicoRequestDTO;
+import com.project.softwave.backend_SoftWave.dto.UsuarioFisico.UsuarioFisicoResponseDTO;
 import com.project.softwave.backend_SoftWave.entity.UsuarioFisico;
 import com.project.softwave.backend_SoftWave.service.UsuarioFisicoService;
 import io.swagger.v3.oas.annotations.Operation;
@@ -30,11 +32,10 @@ public class UsuarioFisicoController {
     })
     @PostMapping
     @SecurityRequirement(name = "Bearer")
-    public ResponseEntity<UsuarioFisicoDTO> cadastrar(@Valid @RequestBody UsuarioFisicoDTO request){
-        UsuarioFisico usuarioFisico = UsuarioFisicoDTO.toEntity(request);
+    public ResponseEntity<UsuarioFisicoResponseDTO> cadastrar(@Valid @RequestBody UsuarioFisicoRequestDTO request){
+        UsuarioFisico usuarioFisico = UsuarioFisicoRequestDTO.toEntity(request);
         UsuarioFisico usuarioNovo = usuarioFisicoService.cadastrar(usuarioFisico);
-
-        return ResponseEntity.status(201).body(UsuarioFisicoDTO.toResponseDto(usuarioNovo));
+        return ResponseEntity.status(201).body(UsuarioFisicoResponseDTO.toResponseDto(usuarioNovo));
     }
 
     @Operation(summary = "Exclusão de um usuário físico", method = "DELETE")
@@ -75,11 +76,11 @@ public class UsuarioFisicoController {
     })
     @GetMapping
     @SecurityRequirement(name = "Bearer")
-    public ResponseEntity<List<UsuarioFisicoDTO>> listar(){
+    public ResponseEntity<List<UsuarioFisicoResponseDTO>> listar(){
         List<UsuarioFisico> usuariosFisicos = usuarioFisicoService.listar();
 
-        List<UsuarioFisicoDTO> usuarioFisicoDtos = usuariosFisicos.stream()
-                .map(UsuarioFisicoDTO::toResponseDto)
+        List<UsuarioFisicoResponseDTO> usuarioFisicoDtos = usuariosFisicos.stream()
+                .map(UsuarioFisicoResponseDTO::toResponseDto)
                 .toList();
 
         return ResponseEntity.status(200).body(usuarioFisicoDtos);
@@ -92,8 +93,8 @@ public class UsuarioFisicoController {
     })
     @GetMapping("/{id}")
     @SecurityRequirement(name = "Bearer")
-    public ResponseEntity<UsuarioFisicoDTO> buscarPorId(@PathVariable Integer id) {
+    public ResponseEntity<UsuarioFisicoResponseDTO> buscarPorId(@PathVariable Integer id) {
         UsuarioFisico usuario = usuarioFisicoService.buscarPorId(id);
-        return ResponseEntity.ok(UsuarioFisicoDTO.toResponseDto(usuario));
+        return ResponseEntity.ok(UsuarioFisicoResponseDTO.toResponseDto(usuario));
     }
 }
