--- src/main/java/com/project/softwave/backend_SoftWave/dto/UsuarioLoginDTO.java
@@ -1,30 +1,0 @@
-package com.project.softwave.backend_SoftWave.dto;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-
-public class UsuarioLoginDTO {
-
-    @Email
-    @NotBlank
-    private String email;
-
-    @NotBlank
-    private String senha;
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getSenha() {
-        return senha;
-    }
-
-    public void setSenha(String senha) {
-        this.senha = senha;
-    }
-}
